server/app/requirements.txt
fastapi
python-dotenv
sqlalchemy
pydantic
passlib
bcrypt
pyjwt